+++ nf_core/pipeline-template/lib/NfcoreTemplate.groovy
@@ -3,6 +3,7 @@
 //
 
 import org.yaml.snakeyaml.Yaml
+import groovy.json.JsonOutput
 
 class NfcoreTemplate {
 
@@ -220,6 +221,21 @@ class NfcoreTemplate {
         if (! postRC.equals(200)) {
             log.warn(post.getErrorStream().getText());
         }
+    }
+
+    //
+    // Dump pipeline parameters in a json file
+    //
+    public static void dump_parameters(params) {
+        def output_d = new File("${params.outdir}/pipeline_info/")
+        if (!output_d.exists()) {
+            output_d.mkdirs()
+        }
+
+        def output_pf  = new File(output_d, "params.json")
+        def jsonStr    = ""
+        jsonStr        = JsonOutput.toJson(params)
+        output_pf.text = JsonOutput.prettyPrint(jsonStr)
     }
 
     //